Move rws client to own package
