+++ abb_rws_client/include/abb_rws_client/rws_state_publisher_ros.hpp
@@ -35,31 +35,44 @@
  */
 
 // This file is a modified copy from
-// https://github.com/ros-industrial/abb_robot_driver/tree/master/abb_rws_service_provider/src
-// https://github.com/ros-industrial/abb_robot_driver/tree/master/abb_rws_state_publisher/src
+// https://github.com/ros-industrial/abb_robot_driver/tree/master/abb_rws_service_provider/include/abb_rws_service_provider/
+// https://github.com/ros-industrial/abb_robot_driver/tree/master/abb_rws_state_publisher/include/abb_rws_state_publisher/
 
-#include "abb_hardware_interface/rws_client.hpp"
+#pragma once
 
-#include "abb_hardware_interface/utilities.hpp"
+#include "abb_rws_client/rws_client.hpp"
+
+// SYSTEM
+#include <string>
+
+// ROS
+#include <rclcpp/rclcpp.hpp>
+
+// ROS INTERFACES
+#include <sensor_msgs/msg/joint_state.hpp>
+
+// ABB INTERFACES
+#include <abb_rapid_sm_addin_msgs/msg/runtime_state.hpp>
+#include <abb_robot_msgs/msg/rapid_task_state.hpp>
+#include <abb_robot_msgs/msg/system_state.hpp>
 
 namespace abb_rws_client {
-RWSClient::RWSClient(const rclcpp::Node::SharedPtr &node, const std::string &robot_ip, unsigned short robot_port)
-    : node_(node),
-      rws_manager_{robot_ip, robot_port, abb::rws::SystemConstants::General::DEFAULT_USERNAME,
-                   abb::rws::SystemConstants::General::DEFAULT_PASSWORD} {
-  node_->declare_parameter("robot_nickname", std::string{});
-  node_->declare_parameter("no_connection_timeout", false);
 
-  connect();
-}
-void RWSClient::connect() {
-  std::string robot_id;
-  bool no_connection_timeout;
+class RWSStatePublisherROS : RWSClient {
+ public:
+  RWSStatePublisherROS(const rclcpp::Node::SharedPtr &node, const std::string &robot_ip, unsigned short robot_port);
 
-  node_->get_parameter("robot_nickname", robot_id);
-  node_->get_parameter("no_connection_timeout", no_connection_timeout);
-  robot_controller_description_ =
-      abb::robot::utilities::establish_rws_connection(rws_manager_, robot_id, no_connection_timeout);
-  abb::robot::utilities::verify_robotware_version(robot_controller_description_.header().robot_ware_version());
-}
+ private:
+  void timer_callback();
+
+  rclcpp::TimerBase::SharedPtr timer_;
+
+  rclcpp::Publisher<sensor_msgs::msg::JointState>::SharedPtr joint_state_pub_;
+  rclcpp::Publisher<abb_robot_msgs::msg::SystemState>::SharedPtr system_state_pub_;
+  rclcpp::Publisher<abb_rapid_sm_addin_msgs::msg::RuntimeState>::SharedPtr runtime_state_pub_;
+
+  abb::robot::MotionData motion_data_;
+  abb::robot::SystemStateData system_state_data_;
+};
+
 }  // namespace abb_rws_client
