+++ abb_rws_client/package.xml
@@ -1,24 +1,19 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>abb_hardware_interface</name>
+  <name>abb_rws_client</name>
   <version>0.0.0</version>
-  <description>ABB EGM driver for ros2 control</description>
-
-  <author email="dmitri.ignakov@taigarobotics.com">Dmitri Ignakov</author>
-
-  <maintainer email="stephanie.eng@picknik.ai">Stephanie Eng</maintainer>
-  <maintainer email="zelenak@picknik.ai">Andy Zelenak</maintainer>
-
+  <description>TODO: Package description</description>
+  <maintainer email="gbartyzel@hotmail.com">Grzegorz Bartyzel</maintainer>
   <license>Apache2</license>
 
-  <buildtool_depend>ament_cmake</buildtool_depend>
-  
-  <depend>abb_egm_rws_managers</depend>
-  <depend>hardware_interface</depend>
-  <depend>pluginlib</depend>
   <depend>rclcpp</depend>
+  <depend>abb_egm_rws_managers</depend>
   <depend>sensor_msgs</depend>
+  <depend>abb_egm_msgs</depend>
+  <depend>abb_rapid_msgs</depend>
+  <depend>abb_robot_msgs</depend>
+  <depend>abb_rapid_sm_addin_msgs</depend>
 
   <test_depend>ament_cmake_gtest</test_depend>
 
@@ -29,3 +24,4 @@
     <build_type>ament_cmake</build_type>
   </export>
 </package>
+
