+++ abb_rws_client/src/rws_client_node.cpp
@@ -1,7 +1,7 @@
 #include <rclcpp/rclcpp.hpp>
 
-#include "abb_hardware_interface/rws_service_provider_ros.hpp"
-#include "abb_hardware_interface/rws_state_publisher_ros.hpp"
+#include "abb_rws_client/rws_service_provider_ros.hpp"
+#include "abb_rws_client/rws_state_publisher_ros.hpp"
 
 int main(int argc, char** argv) {
   rclcpp::init(argc, argv);
