+++ abb_rws_client/src/rws_service_provider_ros.cpp
@@ -38,12 +38,12 @@
 // https://github.com/ros-industrial/abb_robot_driver/tree/master/abb_rws_service_provider/src
 // https://github.com/ros-industrial/abb_robot_driver/tree/master/abb_rws_state_publisher/src
 
-#include "abb_hardware_interface/rws_service_provider_ros.hpp"
+#include "abb_rws_client/rws_service_provider_ros.hpp"
 
 #include <abb_robot_msgs/msg/service_responses.hpp>
 
-#include "abb_hardware_interface/mapping.hpp"
-#include "abb_hardware_interface/utilities.hpp"
+#include "abb_rws_client/mapping.hpp"
+#include "abb_rws_client/utilities.hpp"
 
 using RAPIDSymbols = abb::rws::RWSStateMachineInterface::ResourceIdentifiers::RAPID::Symbols;
