+++ abb_rws_client/src/utilities.cpp
@@ -38,7 +38,7 @@
 // https://github.com/ros-industrial/abb_robot_driver/blob/master/abb_robot_cpp_utilities/src/initialization.cpp
 // https://github.com/ros-industrial/abb_robot_driver/blob/master/abb_robot_cpp_utilities/src/verification.cpp
 
-#include <abb_hardware_interface/utilities.hpp>
+#include <abb_rws_client/utilities.hpp>
 
 #include <stdexcept>
 
@@ -70,7 +70,7 @@ constexpr uint8_t RWS_RECONNECTION_WAIT_TIME{1};
 auto LOGGER = rclcpp::get_logger("ABBHardwareInterfaceUtilities");
 }  // namespace
 
-RobotControllerDescription establishRWSConnection(
+RobotControllerDescription establish_rws_connection(
   RWSManager & rws_manager, const std::string & robot_controller_id,
   const bool no_connection_timeout)
 {
